// src/app/models/modules/contracts/contract.interface.ts
import { IReturnValue } from "../../generals/returnValue.interface";

export interface IContract  extends  IReturnValue{
    hash: string;
    programName: string;
    groupCode: string;
    positionCode: string;
    positionDisplayName: string;
    code: string;
    number: string;
    name: string
    admissionDate: Date;
    datereceipt:Date;
    audCreationDate: Date;
    paymentDate: Date;
    profileName: string;
    mediumCommunicationName: string;
    documentReceptionChargeName: string;
    choiceReturnContributionName: string;
    entryNumber: number;
    term: number;
    impEnrollmentSubTotal: number;
    impEnrollmentTax: number;
    impEnrollmentTotal: number;
    impCapitalFee: number;
    impCapitalAdm: number;
    impCapitalTax: number;
    impCapitalTotal: number;
    impCapitalTotalDisplay: string;
    amountName: string;
    certificateAmount: number;
    certificateAmountDisplay: string;
    associatedNumber: string;
    associatedName: string;
    programHash: string;
    groupHash: string;
    positionHash: string;
    positionEntryCode: string;
    awardedName: string;
    awardedApplyType: string;
    associatedHash: string;
    associatedDocumentTypeHash: string;
    associatedNationalityHash: string;
    documentTypeName: string;
    nationalityName: string;
    displayName: string;
    typeRegisterName: string;
    glcDisplayName: string;
    glcHash: string;
    certificateHash: string;
    profileHash: string;
    contractStatusHash: string;
    mediumCommunicationHash: string;
    documentReceptionChargeHash: string;
    declarationRelationShipHash: string;
    choiceReturnContributionHash: string;
    typeRegisterHash: string;
    currencyName: string;
    storeHash: string;
    storeName: string;
    sellerHash: string;
    sellerName: string;
    subjectHash: string;
    subjectName: string;
    originHash: string;
    originName: string;
    ciGlcCode: string;
    ciAddress: string;
    ciReference: string;
    ibBankHash: string;
    ibBankName: string;
    ibTypeAccountHash: string;
    ibTypeAccountName: string;
    ibAccountNumber: string;
    contractStatusName: string;
    brandHash: string;
    brandName: string;
    modelHash: string;
    modelName: string;
    /*Proposal */
    meetingDate: string;
    meetingInfo: string;
    meetingHash: string;
    hasReturn: string;
    sapStatus: string;
    list: IContract[];
    // ListAssociated: IAssociated[];
    statusRemate: string;

    typeIncomeHash: string;
    typeIncomeName: string;
    feeHash: string;
    accruedNumber: string;
    accruedObservation: string;

    sellerSystem?: number;
    associatedSecond?: string;
}